+++ D2P_GrasshopperTools/D2PGrasshopperToolsInfo.cs
@@ -7,13 +7,13 @@ namespace D2P_GrasshopperTools
     public class D2PGrasshopperToolsInfo : GH_AssemblyInfo
     {
         public override string Name => "d2p-components";
-        public override string Version => "1.0.0";
+        public override string Version => "1.0.2";
 
         //Return a 24x24 pixel bitmap to represent this GHA library.
         public override Bitmap Icon => Properties.Resources.Logo;
 
         //Return a short string describing the purpose of this GHA library.
-        public override string Description => "D2P Components for Grasshopper";
+        public override string Description => "D2P Components streamline parametric modeling in Grasshopper, helping you manage complex data trees, reduce recomputations, and maintain a structured, scalable workflow from concept to fabrication.";
 
         public override Guid Id => new Guid("94B5D1B5-5BC2-4D6F-B5B3-1FC7391668BA");