+++ D2P_GrasshopperTools/GH/Create/GHCreateComponent.cs
@@ -1,11 +1,9 @@
 ﻿using D2P_Core;
-using D2P_Core.Interfaces;
 using Grasshopper.Kernel;
-using Grasshopper.Kernel.Types;
 using Rhino.Geometry;
 using System;
 
-namespace D2P_GrasshopperTools.GH.Components
+namespace D2P_GrasshopperTools.GH.Create
 {
     public class GHCreateComponent : GHComponentPreview
     {
@@ -15,7 +13,7 @@ namespace D2P_GrasshopperTools.GH.Components
         public GHCreateComponent()
           : base("CreateComponent", "Component",
               "Creates a component instance based on a specific type",
-              "D2P", "Components")
+              "D2P", "01 Create")
         {
         }
 
@@ -48,14 +46,14 @@ namespace D2P_GrasshopperTools.GH.Components
             ComponentType componentType = null;
             var name = string.Empty;
             var plane = Plane.Unset;
-            GH_ObjectWrapper parent = null;
+            Component parent = null;
 
             DA.GetData(0, ref componentType);
             DA.GetData(1, ref name);
             DA.GetData(2, ref plane);
             DA.GetData(3, ref parent);
 
-            var parentName = (parent?.Value as IComponent)?.ShortName ?? parent?.Value?.ToString();
+            var parentName = parent?.ShortName ?? parent?.ToString();
             name = string.IsNullOrEmpty(parentName) ? name : $"{parentName}{componentType.Settings.NameDelimiter}{name}";
 
             var component = new Component(componentType, name, plane);