+++ D2P_GrasshopperTools/GH/Create/GHCreateComponentLayerInfo.cs
@@ -3,7 +3,7 @@ using Grasshopper.Kernel;
 using System;
 using System.Drawing;
 
-namespace D2P_GrasshopperTools.GH.Geometry
+namespace D2P_GrasshopperTools.GH.Create
 {
     public class GHCreateComponentLayerInfo : GHComponentBase
     {
@@ -13,14 +13,14 @@ namespace D2P_GrasshopperTools.GH.Geometry
         public GHCreateComponentLayerInfo()
           : base("CreateLayerInfo", "LayerInfo",
               "Creates a layer-information which can be applied to a component together with geometry",
-              "D2P", "Geometry")
+              "D2P", "01 Create")
         {
         }
 
         /// <summary>
         /// Registers all the input parameters for this component.
         /// </summary>
-        protected override void RegisterInputParams(GH_Component.GH_InputParamManager pManager)
+        protected override void RegisterInputParams(GH_InputParamManager pManager)
         {
             pManager.AddTextParameter("LayerName", "N", "The name of the layer below the root-layer of a component", GH_ParamAccess.item);
             pManager.AddColourParameter("LayerColor", "C", "The color of the layer", GH_ParamAccess.item, Color.Black);
@@ -29,7 +29,7 @@ namespace D2P_GrasshopperTools.GH.Geometry
         /// <summary>
         /// Registers all the output parameters for this component.
         /// </summary>
-        protected override void RegisterOutputParams(GH_Component.GH_OutputParamManager pManager)
+        protected override void RegisterOutputParams(GH_OutputParamManager pManager)
         {
             pManager.AddGenericParameter("LayerInfo", "L", "The defined layer-information", GH_ParamAccess.item);
         }
@@ -54,12 +54,12 @@ namespace D2P_GrasshopperTools.GH.Geometry
         /// <summary>
         /// Provides an Icon for the component.
         /// </summary>
-        protected override System.Drawing.Bitmap Icon
+        protected override Bitmap Icon
         {
             get
             {
                 //You can add image files to your project resources and access them like this:                
-                return Properties.Resources.GH_CreateComponentLayerInfo;
+                return Properties.Resources.GH_CreateLayerObject;
             }
         }